minor stuff

+++ src/include/level.h
@@ -24,7 +24,7 @@ class Level : public GameComponent {
 
         bool is_stopped;
 
-        bool position_bound(float pos) const { return in_range<float>(0, pos, lines.size(), true, false); }
+        bool position_bound(float pos) const { return in_range<int>(0, pos, lines.size(), true, false); }
         bool coordinate_bound(float pos) const { return in_range<float>(0, pos, Line::getLineDimension()-pos); }
 
         bool player_in_area() const { return position_bound(player.next_pos()) and coordinate_bound(player.next_coord()); }
